Changed the data type of portnum to CORBA::UShort.

+++ trunk/TAO/tests/Param_Test/options.cpp
@@ -121,7 +121,7 @@ Options::hostname (void)
   return this->hostname_;
 }
 
-CORBA::ULong
+CORBA::UShort
 Options::portnum (void)
 {
   return this->portnum_;
